Atoms / molecules index moved
