+++ src/components/atoms/button/index.ts
@@ -1,0 +1,3 @@
+import Button from "./Button"
+
+export default Button
