+++ src/components/atoms/cartIcon/index.ts
@@ -1,0 +1,3 @@
+import CartIcon from "./CartIcon"
+
+export default CartIcon
