+++ src/components/atoms/index.ts
@@ -1,0 +1,4 @@
+import Button from "./button"
+import CartIcon from "./cartIcon"
+
+export { Button, CartIcon }
