+++ src/components/molecules/cartDropdown/index.ts
@@ -1,0 +1,3 @@
+import CartDropdown from "./CartDropdown"
+
+export default CartDropdown
