+++ src/components/molecules/cartItem/index.ts
@@ -1,0 +1,3 @@
+import CartItem from "./CartItem"
+
+export default CartItem
