+++ src/components/molecules/checkoutItem/index.ts
@@ -1,0 +1,3 @@
+import CheckoutItem from "./CheckoutItem"
+
+export default CheckoutItem
