+++ src/components/molecules/directoryItem/index.ts
@@ -1,0 +1,3 @@
+import DirectoryItem from "./DirectoryItem"
+
+export default DirectoryItem
