+++ src/components/molecules/directoryMenu/index.ts
@@ -1,0 +1,3 @@
+import DirectoryMenu from "./DirectoryMenu"
+
+export default DirectoryMenu
