+++ src/components/molecules/formInput/index.ts
@@ -1,0 +1,3 @@
+import FormInput from "./FormInput"
+
+export default FormInput
