+++ src/components/molecules/productCard/index.ts
@@ -1,0 +1,3 @@
+import ProductCard from "./ProductCard"
+
+export default ProductCard
