+++ src/components/organisms/navigationBar/NavigationBar.tsx
@@ -5,7 +5,7 @@ import { CartContext } from "../../../context/cartContext"
 import { UserContext } from "../../../context/userContext"
 import { signOutUser } from "../../../utils/firebase/FirebaseUtils"
 import CartIcon from "../../atoms/cartIcon/CartIcon"
-import CartDropdown from "../../molecules/cartDropdown/CartDropdown"
+import { CartDropdown } from "../../molecules"
 
 import {
   CenteredLinks,
